pyproject.toml
[project]
name = "uipath-langchain"
version = "0.9.24"
description = "Python SDK that enables developers to build and deploy LangGraph agents to the UiPath Cloud Platform"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.11"
dependencies = [
    "uipath>=2.10.29, <2.11.0",
    "uipath-core>=0.5.2, <0.6.0",
    "uipath-platform>=0.1.24, <0.2.0",
    "uipath-runtime>=0.10.0, <0.11.0",
    "langgraph>=1.0.0, <2.0.0",
    "langchain-core>=1.2.11, <2.0.0",
    "langgraph-checkpoint-sqlite>=3.0.3, <4.0.0",
    "langchain-openai>=1.0.0, <2.0.0",
    "langchain>=1.0.0, <2.0.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "openinference-instrumentation-langchain>=0.1.56",
    "jsonschema-pydantic-converter>=0.4.0",
    "jsonpath-ng>=1.7.0",
    "mcp==1.26.0",
    "langchain-mcp-adapters==0.2.1",
    "pillow>=12.1.1",
]

classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
maintainers = [
    { name = "Marius Cosareanu", email = "marius.cosareanu@uipath.com" },
    { name = "Cristian Pufu", email = "cristian.pufu@uipath.com" },
]

[project.optional-dependencies]
vertex = ["langchain-google-genai>=2.0.0", "google-generativeai>=0.8.0"]
bedrock = ["langchain-aws>=0.2.35", "boto3-stubs>=1.41.4"]
optimize = ["dspy>=2.6"]

[project.entry-points."uipath.middlewares"]
register = "uipath_langchain.middlewares:register_middleware"

[project.entry-points."uipath.runtime.factories"]
langgraph = "uipath_langchain.runtime:register_runtime_factory"

[project.urls]
Homepage = "https://uipath.com"
Repository = "https://github.com/UiPath/uipath-langchain-python"
Documentation = "https://uipath.github.io/uipath-python/"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "ruff>=0.9.4",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-asyncio>=1.0.0",
    "pre-commit>=4.5.1",
    "filelock>=3.20.3",
    "virtualenv>=20.36.1",
    "numpy>=1.24.0",
    "pytest_httpx>=0.35.0",
    "rust-just>=1.39.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/uipath_langchain"]

[tool.ruff]
line-length = 88
indent-width = 4
exclude = ["samples/**", "testcases/**", "template/**"]

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.ruff.lint.per-file-ignores]
"*" = ["E501"]
"src/uipath_langchain/chat/bedrock.py" = ["E402"]
"src/uipath_langchain/chat/vertex.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
plugins = ["pydantic.mypy"]
exclude = ["samples/.*", "testcases/.*", "template/.*"]

namespace_packages = true
follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true

disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-ra -q"
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
